Split table in two
